32 bits support

+++ src/Block.cpp
@@ -8,9 +8,9 @@ Block::Block(const Json::Value &data) : _data(data)
 {}
 
 
-unsigned Block::getIndex() const
+size_t Block::getIndex() const
 {
-    return unhex<uint32_t>(_data["number"].asCString());
+    return unhex<size_t>(_data["number"].asCString());
 }
 
 std::string Block::getHash() const
@@ -23,34 +23,34 @@ std::string Block::getMiner() const
     return _data["miner"].asString();
 }
 
-uint64_t Block::getNonce() const
+size_t Block::getNonce() const
 {
-    return unhex<uint64_t>(_data["nonce"].asCString());
+    return unhex<size_t>(_data["nonce"].asCString());
 }
 
-uint64_t Block::getDifficulty() const
+size_t Block::getDifficulty() const
 {
-    return unhex<uint64_t>(_data["difficulty"].asCString());
+    return unhex<size_t>(_data["difficulty"].asCString());
 }
 
 size_t Block::getSize() const
 {
-    return unhex<uint32_t>(_data["size"].asCString());
+    return unhex<size_t>(_data["size"].asCString());
 }
 
 time_t Block::getTimestamp() const
 {
-    return unhex<uint64_t>(_data["timestamp"].asCString());
+    return unhex<size_t>(_data["timestamp"].asCString());
 }
 
-uint64_t Block::getGasLimit() const
+size_t Block::getGasLimit() const
 {
-    return unhex<uint64_t>(_data["gasLimit"].asCString());
+    return unhex<size_t>(_data["gasLimit"].asCString());
 }
 
-uint64_t Block::getGasUsed() const
+size_t Block::getGasUsed() const
 {
-    return unhex<uint64_t>(_data["gasUsed"].asCString());
+    return unhex<size_t>(_data["gasUsed"].asCString());
 }
 
 Collection<Transaction> Block::getTransactions() const
